Refactor and changes to Character Model and ViewModel
