--- Project/Assets/Editor/EditorModeTests.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: cd5648f7763f38c40b5f0f87e903245a
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
