--- Project/Assets/Editor/EditorModeTests/CharacterModelTests.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a4d90e270a6ed694dbbe80878aa73e4b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
