--- Project/Assets/Editor/EditorModeTests/EditorModeTests.asmdef
@@ -1,14 +1,0 @@
-{
-    "name": "EditorModeTests",
-    "references": [
-        "ScriptsAssembly"
-    ],
-    "optionalUnityReferences": [
-        "TestAssemblies"
-    ],
-    "includePlatforms": [
-        "Editor"
-    ],
-    "excludePlatforms": [],
-    "allowUnsafeCode": false
-}
