--- Project/Assets/Editor/EditorModeTests/EditorModeTests.asmdef.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: cf30041875a574848bcf54567d04ad4b
-AssemblyDefinitionImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
