--- Project/Assets/Editor/EditorModeTests/character.xml.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: cc4bee65206ffc84fbabb6808f56b802
-TextScriptImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
