--- Project/Assets/PlayModeTests.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 955b4f1c7bf6c0148b7a10953a0e53e8
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
