--- Project/Assets/PlayModeTests/NewTestScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f9b60d4fe0e2e5e40a9d3ff70dce1761
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
