--- Project/Assets/PlayModeTests/PlayModeTests.asmdef
@@ -1,12 +1,0 @@
-{
-    "name": "PlayModeTests",
-    "references": [
-        "ScriptsAssembly"
-    ],
-    "optionalUnityReferences": [
-        "TestAssemblies"
-    ],
-    "includePlatforms": [],
-    "excludePlatforms": [],
-    "allowUnsafeCode": false
-}
