--- Project/Assets/PlayModeTests/PlayModeTests.asmdef.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 00edeacbc7802be448d39840d7de6ea1
-AssemblyDefinitionImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
