--- Project/Assets/ScriptableObjects/Character.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 2579dd979384f412cb4e0a465ad86a5c
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
