--- Project/Assets/Scripts/Player Code.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 704508a7b91689f4d80bd66b7ea37e51
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
