--- Project/Assets/Scripts/Player Code/InputControllers.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d603fe8514b214e6995d1f60fd164de0
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
