--- Project/Assets/Scripts/Player Code/InputControllers/Platform dependent.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 434d1848daa9e4843b8185e534ff3252
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
