--- Project/Assets/Scripts/Utility.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: aedc1b95e7bc2b14889c6820dbeacc17
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
