+++ Project/Assets/_Scenes/HealthTest.unity.meta
@@ -1,6 +1,5 @@
 fileFormatVersion: 2
-guid: 47d54edb71f0f144a8d97aa9e9792e13
-folderAsset: yes
+guid: afd0301235f18764da953a21aac6b779
 DefaultImporter:
   externalObjects: {}
   userData:
