--- Project/Assets/_Scripts/_new_Scripts/Player Code/CharacterHealthModel.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1aaa41478e54840f4a79dc6b74d7f3b0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
