+++ Project/Assets/_Scripts/_new_Scripts/Player Code/CharacterViewModel.cs
@@ -3,27 +3,48 @@
 using UnityEngine;
 
 using Steamy.Player.MotionModes;
+using UnityEngine.UI;
+using System.ComponentModel;
 
 namespace Steamy.Player
 {
 	public class CharacterViewModel : MonoBehaviour
 	{
-		[SerializeField]
 		public CharacterModel Model;
 
-		public string XmlFilePath;
+        public Text HealthView;
 
 		private void Awake()
 		{
-			Model = CharacterModel.ReadFromXml(XmlFilePath, new Type[] { typeof(MotionMode) });
-			Model.ViewModel = this;
-		}
+            Model.Health.PropertyChanged += OnHealthChanged;
+            Model.Health.Value = Model.Health.MaxValue;
+        }
 
-
-		private void Update()
+        private void Update()
 		{
-			// Move the character based on player input
-			Model.DoMotion();
+
+            MoveCharacter();
+
+            if (Input.GetKeyDown(KeyCode.K))
+            {
+                Model.Health.Value -= 1;
+            }
 		}
-	}
+
+        private void MoveCharacter()
+        {
+            if (Model.MotionModes == null)
+                return;
+
+            foreach (var mode in Model.MotionModes)
+            {
+                mode.Do(this);
+            }
+        }
+
+        private void OnHealthChanged(object sender, PropertyChangedEventArgs e)
+        {
+            HealthView.text = Model.Health.Value.ToString();
+        }
+    }
 }
